Fix canEdit for admins (Fix #9474)
This was a good deal more complicated than would might
have expected. In order to handle the group==-1 case,
it was necessary to pass in the details to isGraphCritical.
However, to be able to simultaneously check each of the
different types of writes (ANNOTATE, EDIT, DELETE, LINK)
it was necessary to re-work allowUpdateOrDelete to check
multiple at the same time.